# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "InfraPulse MSK",
    "description": "API InfraPulse MSK. Документация открыта для чтения; каждый маршрут, кроме `/health/*` и входа, проверяет сессию или токен интеграции.\n\n* **Чтение (GET)** — по cookie сессии: войти в веб-интерфейс этого же сайта, затем «Try it out» здесь.\n* **Пачки наблюдений** `POST /api/v1/observations` — токен интеграции: «Authorize» → `integrationToken`, значение `ipk_...` без слова Bearer.\n* **Запись под сессией (POST)** — дополнительно `csrfToken`: значение cookie `__Host-infrapulse-csrf`.\n\nИнструкция для систем-источников — `docs/INTEGRATION_API.md` в репозитории.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/observations": {
      "post": {
        "summary": "Ingest",
        "operationId": "ingest_api_v1_observations_post",
        "requestBody": {
          "description": "ObservationBatch as JSON, or the same batch in XML: root <batch> with the batch_id attribute (or a <batch_id> element) and <record> elements whose child elements are the record fields; alarm is true or false. DTD and entities are refused (422 xml_forbidden). docs/INTEGRATION_API.md",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObservationBatch"
              },
              "example": {
                "batch_id": "scada-ods-1-20260630T120000-000184",
                "records": [
                  {
                    "event_id": "9000001",
                    "channel_id": "demo-700001",
                    "date": "2026-06-30",
                    "time": "11:59:58",
                    "alarm": false,
                    "value": "28"
                  },
                  {
                    "ид_события": "9000002",
                    "ид_канала_данных": "demo-700002",
                    "дата": "2026-06-30",
                    "время": "11:59:59",
                    "тревожное": false,
                    "значение_датчика": "Неисправен"
                  },
                  {
                    "event_id": "9000003",
                    "channel_id": "demo-700003",
                    "event_at": "2026-06-30T09:00:00Z",
                    "alarm": false,
                    "value": "Обрыв; ТО"
                  }
                ]
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/ObservationBatch"
              },
              "example": "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<batch batch_id=\"scada-ods-1-20260630T120000-000184\">\n  <record>\n    <event_id>9000001</event_id>\n    <channel_id>demo-700001</channel_id>\n    <date>2026-06-30</date>\n    <time>11:59:58</time>\n    <alarm>false</alarm>\n    <value>28</value>\n  </record>\n  <record>\n    <ид_события>9000002</ид_события>\n    <ид_канала_данных>demo-700002</ид_канала_данных>\n    <дата>2026-06-30</дата>\n    <время>11:59:59</время>\n    <тревожное>false</тревожное>\n    <значение_датчика>Неисправен</значение_датчика>\n  </record>\n  <record>\n    <event_id>9000003</event_id>\n    <channel_id>demo-700003</channel_id>\n    <event_at>2026-06-30T09:00:00Z</event_at>\n    <alarm>false</alarm>\n    <value>Обрыв; ТО</value>\n  </record>\n</batch>\n"
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportFile"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type is not JSON or XML"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "integrationToken": []
          },
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ]
      }
    },
    "/api/v1/observations/{import_id}": {
      "get": {
        "summary": "Batch Status",
        "operationId": "batch_status_api_v1_observations__import_id__get",
        "parameters": [
          {
            "name": "import_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Import Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportFile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "integrationToken": []
          },
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportFile": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Id"
          },
          "format": {
            "type": "string",
            "enum": [
              "journal_csv",
              "journal_json",
              "reference_channels_csv",
              "reference_objects_csv",
              "reference_states_csv"
            ],
            "title": "Format"
          },
          "file_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "File Name"
          },
          "sha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256"
          },
          "size_bytes": {
            "type": "integer",
            "maximum": 52428800.0,
            "minimum": 0.0,
            "title": "Size Bytes"
          },
          "uploaded_by": {
            "type": "string",
            "minLength": 1,
            "title": "Uploaded By"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Uploaded At"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "parsing",
              "imported",
              "recomputing",
              "published",
              "duplicate",
              "failed"
            ],
            "title": "Status"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "rows_total": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows Total"
          },
          "rows_accepted": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows Accepted"
          },
          "rows_duplicate": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows Duplicate"
          },
          "rows_quarantined": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows Quarantined"
          },
          "unknown_channels": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Unknown Channels"
          },
          "event_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event From"
          },
          "event_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event To"
          },
          "reference_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Version"
          },
          "duplicate_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duplicate Of"
          },
          "forecast_generation": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Generation"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "file_too_large",
                  "unknown_format",
                  "bad_header",
                  "bad_encoding",
                  "no_valid_rows",
                  "reference_missing",
                  "recompute_failed",
                  "internal_error"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "timings": {
            "items": {
              "$ref": "#/components/schemas/ImportStageTiming"
            },
            "type": "array",
            "title": "Timings"
          },
          "quarantine_reasons": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "enum": [
                "bad_column_count",
                "bad_date",
                "bad_time",
                "bad_bool",
                "empty_channel",
                "bad_encoding",
                "value_too_long",
                "channel_type_conflict"
              ]
            },
            "type": "object",
            "title": "Quarantine Reasons"
          },
          "quarantine_sample": {
            "items": {
              "$ref": "#/components/schemas/QuarantineSample"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Quarantine Sample"
          },
          "new_card_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "New Card Ids"
          },
          "released_card_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Released Card Ids"
          },
          "source_layout": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "organizers_export",
                  "tz_appendix1",
                  "api_batch",
                  "reference"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Layout"
          },
          "source_container": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "csv",
                  "xlsx",
                  "json",
                  "xml"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Container"
          },
          "alarm_not_provided": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Alarm Not Provided"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Notes"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "format",
          "file_name",
          "sha256",
          "size_bytes",
          "uploaded_by",
          "uploaded_at",
          "status"
        ],
        "title": "ImportFile",
        "description": "One uploaded file and its processing report.\n\nRow counts: ``rows_total = rows_accepted + rows_duplicate + rows_quarantined``.\n``unknown_channels`` are channels absent from the reference; their rows are kept\nwith an unknown object (DATA-04). ``forecast_generation`` is the publication that\nfirst included the file."
      },
...
      "ObservationBatch": {
        "properties": {
          "batch_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._:-]+$",
            "title": "Batch Id"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/ObservationRecord"
            },
            "type": "array",
            "maxItems": 5000,
            "minItems": 1,
            "title": "Records"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "batch_id",
          "records"
        ],
        "title": "ObservationBatch",
        "description": "A batch of records from an integration; ``batch_id`` is the client's idempotency key.\n\nThe same ``batch_id`` with the same bytes returns the stored import; with other bytes\nit is a conflict (409). Records are also deduplicated with the CSV import key.\nThe body is JSON or the same batch in XML (``<batch batch_id=\"...\">`` with\n``<record>`` elements whose child elements are these fields); both are validated\nby this model."
      },
      "ObservationRecord": {
        "properties": {
          "event_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Event Id"
          },
          "channel_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Channel Id"
          },
          "date": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "time": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{2}:\\d{2}:\\d{2}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time"
          },
          "event_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event At"
          },
          "alarm": {
            "type": "boolean",
            "title": "Alarm"
          },
          "value": {
            "type": "string",
            "maxLength": 1000,
            "title": "Value"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event_id",
          "channel_id",
          "alarm",
          "value"
        ],
        "title": "ObservationRecord",
        "description": "One journal record with the fields of the CSV export.\n\nKeys may be given in English or as in the CSV header: ``event_id`` / ``ид_события``,\n``channel_id`` / ``ид_канала_данных``, ``alarm`` / ``тревожное``, ``value`` /\n``значение_датчика``. Time is either ``date`` + ``time`` (local Moscow time, the same\nrule as the CSV: ``дата`` ``YYYY-MM-DD`` and ``время`` ``HH:MM:SS``) or ``event_at``\n(ISO 8601 with an offset), never both. ``value`` is the source text kept verbatim."
      },
...
    },
    "securitySchemes": {
      "integrationToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ipk_...",
        "description": "Токен интеграции внешней системы (выдаёт администратор сервиса). Даёт только право ingest: отправку пачек и статус своих пачек."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-infrapulse-session",
        "description": "Сессия входа через каталог. Ставится при входе в веб-интерфейс этого же сайта и отправляется браузером сама; поле в «Authorize» заполнять не нужно."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Для POST под сессией: значение cookie __Host-infrapulse-csrf. С токеном интеграции не нужен."
      }
    }
  }
}